refactor(mybookkeeper/leases): extract inline components + move ORM out of service
Pre-commit review feedback:

- Move ``db.flush()`` out of ``lease_template_service.update_placeholder``
  into the new ``lease_template_placeholder_repo.update_fields`` helper.
  Services should orchestrate (load → validate → call repo); the repo owns
  ORM mutations (per the layered-architecture rule).
- Extract ``PlaceholderSpecRow`` from ``PlaceholderSpecEditor.tsx`` into its
  own file. One component per file.
- Extract ``PlaceholderInput`` from ``LeaseGenerateForm.tsx`` into its own
  file. One component per file.
- Replace the two layout-only E2E tests with two real flow tests: one that
  edits a placeholder's display label and verifies the change persists in
  the DB, and one that deletes a template via the UI and verifies the
  soft-delete via the API. Every test now creates → acts → verifies →
  cleans up.

No behaviour changes. All 37 backend lease tests still pass; frontend tsc +
vitest still green.
